app/app/auth/user/api/user_router.py
from django.urls import path
from auth.user.api.user_views import UserAddView, UserShowView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('auth/register', UserAddView.as_view()),
    path('auth/login', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('auth/token/refresh', TokenRefreshView.as_view(), name='token_refresh'),
    path('auth/usuario', UserShowView.as_view()),
]