requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.13
django-filter @ git+https://github.com/carltongibson/django-filter.git@af9f13b6245cba71b3da358bd014159474ebfb8c
djangorestframework==3.14.0
filelock==3.12.2
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.1
platformdirs==3.5.3
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.23.0
